requirements.txt
numpy
pandas
sklearn
neat-python
torch
torchvision
torchaudio
matplotlib
torch_geometric
ydata-synthetic